+++ src/main/webapp/resources/js/components/project-user-groups/ProjectUserGroupsTable.jsx
@@ -1,5 +1,5 @@
 import React, { useContext } from "react";
-import { Button, notification } from "antd";
+import { Button } from "antd";
 import { PagedTable, PagedTableContext } from "../ant.design/PagedTable";
 import { setBaseUrl } from "../../utilities/url-utilities";
 import { formatInternationalizedDateTime } from "../../utilities/date-utilities";