+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Changes
 * [UI/Developer]: Removed most dependency for `momentjs` (only on project samples filters).
 * [Developer]: Removed deprecated Long constructor.
 * [Developer]: Updated `webpack-dev-server` to fix security vulnerability caused by `websocket-extensions` 0.1.3.
+* [UI]: Update project metadata templates listing page to use Ant Design.
 * [UI]: Allowing users to synchronize assembly data from remote APIs.
 
 20.01 to 20.05